Prevent rendering quiz section on the server
Since we need access to localStorage for correctly rendering the quiz
while taking into account the user's responses, we just skip rendering
it on the server. We could have done a partial render on the server
and then enhanced it on the client, but we don't have the time to fight
the intricacies of React's hydration rules right now.