Share using valid HTML IDs
History of the Worldcup cards have HTML IDs that start
with a number (the year), which is invalid. As a result,
the printer service is unable to extract the specific
element being shared and errors out.

This change generates a new HTML ID based on the fact
id with a leading underscore to ensure it is valid.

Bug: https://trello.com/c/SzW9u9gG/4-share-feature-shares-a-broken-url