Don't fail when a fact card image is not found
The FactCard component was already trying to handle a fact
that has no associated image but resolving the URL
(a few lines up) was breaking. I think this change
restores the intent of the code when there is no image.
We cannot really use a placeholder image in this situation
since this a generic component that is reused for players,
venues, events, etc.

Also had placeholder images for Rashid Khan.

Bug: https://trello.com/c/rlgSoRJK/17-500-error-on-page-https-cricketinfoio-read-en-rashidkhan